Fix cling when built for 32 bit architectures.
Use uintptr_t as void* and uint64_t are not guaranteed to be the same size.

+++ include/cling/Utils/AST.h
@@ -116,7 +116,7 @@ namespace utils {
     /// cast to.
     ///
     clang::Expr* CStyleCastPtrExpr(clang::Sema* S, clang::QualType Ty,
-                                   uint64_t Ptr);
+                                   uintptr_t Ptr);
 
     ///\brief Synthesizes c-style cast in the AST from given pointer and type to
     /// cast to.
@@ -128,7 +128,7 @@ namespace utils {
     ///
     //  TODO: Use Sema::ActOnIntegerConstant.
     clang::IntegerLiteral* IntegerLiteralExpr(clang::ASTContext& C,
-                                              uint64_t Ptr);
+                                              uintptr_t Ptr);
 
   }
